+++ components/Signup/index.ts
@@ -1,0 +1,4 @@
+import SignupLoading from './SignupLoading/SignupLoading';
+import SignUpModal from './SignUpModal/SignUpModal';
+
+export { SignupLoading, SignUpModal };
