--- components/signup/Loading.module.scss
@@ -1,17 +1,0 @@
-@import '../../styles/mixin';
-
-.container {
-    @include flex-center-col;
-    position: absolute;
-    top : 50%;
-    left : 50%;
-    z-index: 1;;
-    transform: translate(-50%, -50%);
-    background-color: white;
-    padding: 5rem 7rem;
-    border-radius: 2rem;
-
-    h3 {
-        font-size: 1.2rem;
-    }
-}
